pyproject.toml
[tool.poetry]
name = "carpet-market"
version = "0.1.0"
description = ""
authors = ["Сократ"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
sqlalchemy = {version = "^2.0.36", extras = ["asyncio"]}
alembic = "^1.14.0"
pydantic = "^2.10.3"
dishka = "^1.4.2"
environs = "^11.2.1"
uvicorn = "^0.32.1"
asyncpg = "^0.30.0"
python-multipart = "^0.0.20"
boto3 = "^1.36.2"
python-jose = "^3.3.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"


[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

markers = [
    "domain: test related to domain layer",
    "value_objects: test related to value objects"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
